refactor(l5): move namespace-policy and permissions from src/ into server/security (issue 864)
L5 of _plans/server-hardening-ladder.md: cut server/ free of src/.

Moved (git mv, old -> new):
- src/namespace-policy.ts -> server/security/namespace-policy.ts
- src/permissions.ts -> server/security/permissions.ts
- src/permissions.test.ts -> server/security/permissions.test.ts

Runtime closure of src/ under server/main.ts: 46 -> 44.

Namespace isolation is a security boundary. Every predicate in
canWriteNamespace, isPromoterIdentity, writableNamespaces, and
appendWriteNamespacePredicate is byte-identical apart from formatting;
no reason string, no comparison, and no evaluation order changed.

The one structural edit is forced by the server/ standard: canWriteNamespace
scored complexity 14 against the server maximum of 10, so the four
namespace-shaped rejections that precede any role decision (legacy shared
write, frozen snapshot, X-Namespace header mismatch, shared-namespace
promoter identity) are hoisted verbatim into namespaceWriteRejection, which
returns the same NamespaceCheck object or undefined. canWriteNamespace calls
it first and returns its result unchanged, so the guards still run in their
original order ahead of the unchanged role ladder.

The old src/ paths become shims per M2 so that src/ and scripts/ importers
keep their import lines and stay out of the staged lint set. server/
importers already resolve the pre-existing server/auth twins and are
untouched: no retarget.

src/namespace-policy.test.ts (298 lines) is left at its src/ path per the M1
exception and keeps testing the moved code through the shim.

No process.env is read by either moved module.
