+++ src/namespace-policy.ts
@@ -1,173 +1,2 @@
-import type { AuthInfo } from "./types.ts";
-import {
-  physicalNamespace,
-  sharedNamespaceConfig,
-  shouldRejectLegacySharedWrite,
-} from "./shared-namespace.ts";
-
-export const PROMOTER_CLIENT_IDS = new Set([
-  "openbrain-promoter",
-  "hermes-promoter",
-]);
-
-/**
- * Namespaces frozen as read-only snapshots (#167). With `collab` retired as
- * the default legacy shared namespace, nothing would otherwise stop an
- * admin/n8n/promoter token from writing a namespace literally named `collab`
- * and silently breaking the frozen-snapshot model with invisible orphans.
- * This rejection is independent of the legacy-shared mechanism. The only
- * bypass is the explicit operator escape hatch: configuring the namespace as
- * legacy (SHARED_NAMESPACE_LEGACY) with OPENBRAIN_ALLOW_LEGACY_SHARED_WRITES=1
- * for a migration window.
- */
-export const FROZEN_NAMESPACES = new Set(["collab"]);
-
-function isFrozenNamespaceWrite(targetNamespace: string): boolean {
-  if (!FROZEN_NAMESPACES.has(targetNamespace)) return false;
-  const config = sharedNamespaceConfig();
-  // Operator escape hatch: an explicitly configured legacy namespace with
-  // legacy writes enabled stays writable for the migration window.
-  if (
-    config.allowLegacySharedWrites &&
-    config.legacySharedNamespace === targetNamespace
-  ) {
-    return false;
-  }
-  return true;
-}
-
-export interface NamespaceCheck {
-  allowed: boolean;
-  reason?: string;
-}
-
-export function isPromoterIdentity(auth: AuthInfo): boolean {
-  // First-class promoter role (#147) is a promoter identity by role alone.
-  if (auth.role === "promoter") {
-    return true;
-  }
-  // Backward-compat: the legacy clientId-on-admin/ob-admin convention (e.g. the
-  // in-process legacy-shared promoter CLI) remains a promoter identity.
-  return (
-    (auth.role === "admin" || auth.role === "ob-admin") &&
-    PROMOTER_CLIENT_IDS.has(auth.tokenClientId ?? auth.clientId)
-  );
-}
-
-export function canWriteNamespace(
-  auth: AuthInfo,
-  targetNamespace: string,
-): NamespaceCheck {
-  const physicalTargetNamespace = physicalNamespace(targetNamespace);
-  if (shouldRejectLegacySharedWrite(auth, targetNamespace)) {
-    return {
-      allowed: false,
-      reason: `legacy shared namespace '${targetNamespace}' is read-only for normal clients; use '${sharedNamespaceConfig().canonicalSharedNamespace}'`,
-    };
-  }
-
-  // Frozen snapshot namespaces reject writes for ALL roles, including
-  // admin/n8n/promoter (#167). See FROZEN_NAMESPACES.
-  if (
-    isFrozenNamespaceWrite(targetNamespace) ||
-    isFrozenNamespaceWrite(physicalTargetNamespace)
-  ) {
-    return {
-      allowed: false,
-      reason: `namespace '${targetNamespace}' is a frozen snapshot (retired legacy namespace) and rejects all writes`,
-    };
-  }
-
-  if (
-    auth.namespaceSource === "header" &&
-    physicalTargetNamespace !== auth.clientId
-  ) {
-    return {
-      allowed: false,
-      reason: `X-Namespace header requires writes to namespace '${auth.clientId}'`,
-    };
-  }
-
-  const config = sharedNamespaceConfig();
-  if (
-    physicalTargetNamespace === config.physicalSharedNamespace &&
-    !isPromoterIdentity(auth)
-  ) {
-    return {
-      allowed: false,
-      reason:
-        `${config.sharedNamespace} writes require the openbrain-promoter or hermes-promoter service identity`,
-    };
-  }
-
-  if (
-    auth.role === "admin" ||
-    auth.role === "ob-admin" ||
-    auth.role === "promoter"
-  ) {
-    return { allowed: true };
-  }
-
-  if (auth.role === "readonly") {
-    return { allowed: false, reason: "readonly role cannot write" };
-  }
-
-  if (physicalTargetNamespace === auth.clientId) {
-    return { allowed: true };
-  }
-
-  if (auth.role === "discord") {
-    return {
-      allowed: false,
-      reason: `discord role can only write to own namespace (${auth.clientId})`,
-    };
-  }
-
-  return {
-    allowed: false,
-    reason: `${auth.role} role cannot write to namespace '${targetNamespace}'`,
-  };
-}
-
-export function writableNamespaces(auth: AuthInfo): string[] | undefined {
-  if (auth.namespaceSource === "header") {
-    return [auth.clientId];
-  }
-
-  // Promoter writes across namespaces (incl. shared-kb) by design (#147).
-  if (
-    auth.role === "admin" ||
-    auth.role === "ob-admin" ||
-    auth.role === "promoter"
-  ) {
-    return undefined;
-  }
-
-  return [auth.clientId];
-}
-
-export function appendWriteNamespacePredicate(
-  auth: AuthInfo,
-  params: unknown[],
-  column = "namespace",
-): string {
-  const namespaces = writableNamespaces(auth);
-  if (namespaces === undefined) {
-    const excludedNamespaces = Array.from(FROZEN_NAMESPACES).filter(
-      (namespace) => isFrozenNamespaceWrite(namespace),
-    );
-    if (
-      (auth.role === "admin" || auth.role === "ob-admin") &&
-      !isPromoterIdentity(auth)
-    ) {
-      excludedNamespaces.push(sharedNamespaceConfig().sharedNamespace);
-    }
-    if (excludedNamespaces.length > 0) {
-      params.push(excludedNamespaces);
-      return ` AND ${column} <> ALL($${params.length}::text[])`;
-    }
-    return "";
-  }
-  params.push(namespaces);
-  return ` AND ${column} = ANY($${params.length}::text[])`;
-}
+// L5 shim (issue 864): moved to server/security/namespace-policy.ts; retired with src/ at L6.
+export * from "../server/security/namespace-policy.ts";
